chore: move use case modal to separate component
